move utility classes to dedicated physicsutils/ folder
