+++ PhysicsTools/Heppy/python/physicsutils/genutils.py
